+++ cluster/src/main/java/io/atomix/cluster/messaging/impl/ProtocolRequest.java
@@ -26,7 +26,7 @@ public final class ProtocolRequest extends ProtocolMessage {
     private final Address sender;
     private final String subject;
 
-    public ProtocolRequest(int id, Address sender, String subject, byte[] payload) {
+    public ProtocolRequest(long id, Address sender, String subject, byte[] payload) {
         super(id, payload);
         this.sender = sender;
         this.subject = subject;
